Changed some IntegerFields to SmallIntegerFields

+++ api/src/telcameras_v2/models.py
@@ -11,11 +11,11 @@ class Observation(models.Model):
     purpose = postgres_fields.ArrayField(models.CharField(max_length=255), null=True)  # Doel van de camera
     latitude = models.DecimalField(max_digits=9, decimal_places=6)
     longitude = models.DecimalField(max_digits=9, decimal_places=6)
-    interval = models.IntegerField()                        # e.g. 60     # seconds that this message spans
+    interval = models.SmallIntegerField()                   # e.g. 60     # seconds that this message spans
     timestamp_message = models.DateTimeField()
     timestamp_start = models.DateTimeField()
-    # message = models.IntegerField()                         # Volgnummer bericht
-    # message_type = models.CharField(max_length=50)        # either "count" OR "person"
+    # message = models.IntegerField()                       # Volgnummer bericht. Is saved in the aggregates
+    # message_type = models.CharField(max_length=50)        # either "count" OR "person". Is not stored, but used to determine which aggregate to use
 
 
 class CountAggregate(models.Model):
@@ -27,14 +27,14 @@ class CountAggregate(models.Model):
     type = models.CharField(max_length=255)         # e.g. "line" or "zone". No idea what else it can be.
 
     # For a type "line"
-    azimuth = models.IntegerField(null=True)
-    count_in = models.IntegerField(null=True)
-    count_out = models.IntegerField(null=True)
+    azimuth = models.PositiveSmallIntegerField(null=True)
+    count_in = models.SmallIntegerField(null=True)
+    count_out = models.SmallIntegerField(null=True)
 
     # for a type "zone"
     area = models.FloatField(null=True)
     geom = models.TextField(null=True)
-    count = models.IntegerField(null=True)
+    count = models.SmallIntegerField(null=True)
 
 
 class PersonAggregate(models.Model):
